refactor(skills): 将跨文件共享类型集中到 types/
